Refactor namespace hooks: remove old hooks and introduce new hooks to retrieve namespace from query parameters
